requirements.txt
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
groq==0.5.0
joblib==1.3.2
scikit-learn==1.4.1.post1
pandas==2.2.1
chromadb==0.4.24
langchain==0.1.16
sentence-transformers==2.6.1
faiss-cpu==1.8.0
python-dotenv==1.0.0
requests==2.31.0
SpeechRecognition==3.10.1
PyAudio==0.2.14
gTTS==2.5.1
numpy==1.26.4
fastapi==0.110.0
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
sqlalchemy==2.0.29
google-auth==2.29.0